polybase cleanup and plotting module
